requirements.txt
cffi==1.14.4
ExifRead==2.3.2
ffmpeg-python==0.2.0
future==0.18.2
pycparser==2.20
watchdog==1.0.2
